// frontend/src/app/core/api/model/auditLogResponse.ts
/**
 * Rev-PasswordManager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AuditLogResponse { 
    id?: number;
    action?: string;
    details?: string;
    ipAddress?: string;
    timestamp?: string;
}